🔷 Player - Gun 분리 및 컴포넌트 설계 중

+++ Assets/Scripts/Gameplay_System/Player/Stat/SimpleValue.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [Serializable]
 public class SimpleValue
 {
-    [SerializeField] private int value;
+    [SerializeField] private float value;
 
-    public int GetValue() => value;
+    public float GetValue() => value;
 }
